services/auth-service/requirements.txt
fastapi==0.95.1
httpx==0.24.1
python-jose==3.3.0
uvicorn==0.23.1
python-dotenv==1.0.0
python-multipart==0.0.5
bcrypt==3.2.0
